Remove requirements.py
